requirements.txt
bcrypt
streamlit==1.29.0
streamlit-option-menu==0.3.6
pytz
Pillow
psycopg2-binary==2.9.10
matplotlib
pandas
plotly.express